requirements.txt
numpy<2.0 
mujoco
roboticstoolbox-python
ompl
open3d